move SQLExecutor to sharding-core-executor module
