+++ sharding-core/sharding-core-execute/src/main/java/org/apache/shardingsphere/sharding/execute/sql/execute/SQLExecutor.java
@@ -15,11 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.shardingjdbc.executor;
+package org.apache.shardingsphere.sharding.execute.sql.execute;
 
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.sharding.execute.sql.execute.SQLExecuteTemplate;
-import org.apache.shardingsphere.sharding.execute.sql.execute.SQLExecutorCallback;
 import org.apache.shardingsphere.underlying.executor.StatementExecuteUnit;
 import org.apache.shardingsphere.underlying.executor.kernel.InputGroup;
